rename swatch to current color
