+++ .prettierignore
@@ -2,3 +2,4 @@ dist
 node_modules
 .git
 .tsup
+storybook-static
