+++ src/index.tsx
@@ -1,5 +1,5 @@
 import "./index.css";
 
 export * from "@/alpha-slider";
+export * from "@/current-color";
 export * from "@/hue-slider";
-export * from "@/swatch";
